+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove the `deliveryPromisesEnabled` flag to enable delivery promise filters
+
 ## [3.138.1] - 2025-07-14
 
 ### Fix